Refactor: 코드리뷰 반영 #14
